RESUME PROJECT AFTER MONTHS
After months of inactivity, I'm taking this project up again. I left it pending before, so I'm not committing all the files that were changed but with untracked changes.
In the next couple of commits I will also add useful files that weren't committed.

If something has gone wrong, a backup taken today can be restored. It's called:
  PySoundTrap-20160114.tgz
and at the time of writing it can be found at HDDav2/BACKUPs/PySoundTrap-20160114.tgz (no guarantee it will be there in the future).